Extract auxiliary methods to utils file
